BLASTP CD-HIT Input File
The CD-HIT software required an input file to be in a specific 'dataset' format. Poor documentation did not specify what file type this is or how it should be created, so I manually created the desired file type myself using this script. 

(The same commands for BLASTN CD-HIT's input file as well.)